+++ schemas/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+-   Add new properties `settings.spacing.fluid` and `settings.spacing.spacingSizes[n].fluid` to theme.json to enable fluid spacing ([#81864](https://github.com/WordPress/gutenberg/issues/81864)).
 -   Add new properties `settings.typography.fluid` and `settings.typography.fontSizes[n].fluidSize` to theme.json to enable fluid typography ([#39529](https://github.com/WordPress/gutenberg/pull/39529)).